requirements.txt
torch==2.2.2
numpy==1.26.4
pyyaml==6.0.1
ollama==0.6.1
matplotlib==3.10.8
pandas==3.0.2
